// src/generated/RelocateEntrances.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: RelocateEntrances
// ====================================================

export interface RelocateEntrances_updateEntrances {
  id: string;
}

export interface RelocateEntrances {
  updateEntrances: RelocateEntrances_updateEntrances[];
}

export interface RelocateEntrancesVariables {
  ids: string[];
  to: string;
}