// src/generated/DeleteEntrances.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { DeleteManyInput } from "./globalTypes";

// ====================================================
// GraphQL mutation operation: DeleteEntrances
// ====================================================

export interface DeleteEntrances {
  deleteEntrances: boolean;
}

export interface DeleteEntrancesVariables {
  input: DeleteManyInput;
}